Reviewed
(cherry picked from commit 0bf0af292eeb60e1f477351ee213366f00a19233)

(and then merge conflicts solved -- RM)

+++ DFNs/EdgeDetection/CannyDetector.hpp
@@ -44,9 +44,9 @@ namespace dfn_ci
 
 			struct CannyParameters
 			{
-				float lowThreshold;
-				float highThreshold;
-				float kernelSize;
+				double lowThreshold;
+				double highThreshold;
+				int kernelSize;
 			};
 
 			struct CannyDetectorOptionsSet
